ROI with each coord showing # subjects sharing it
This is the R script that will create a final label file in which every coordinate is numbered accordingly with how many subjects share that coordinate in their ROI. This assumes that the label files piped into this script are all sampled to fsaverage space, or any type of common space. The coloring can be adjusted in freesurfer GUI so that every value (# subjects sharing that coordi) shows up as a different color.